Stops importing metal modules with full path
This is important so that soy projects can be precompiled to commonjs for example, while not forcing dependencies to be required from `src` (they'll respect the "main" field from package.json instead).

This is being released as a major version so that projects relying on the full path for some reason aren't affected immediately.